Zustand (#27)
* add fake db, add zustand

* add new answer feature with zustand

* trouble shooting re-render

* bug fix - wrong function name

* improve addAnswer functionality
