+++ src/components/QuestionComponent/index.js
@@ -1,11 +1,31 @@
+import {nanoid} from 'nanoid';
 import {useState} from 'react';
 
+import {useStore} from '../../hooks/useStore';
 import Form from '../Form';
 
 import StyledQuestionComponent from './styled';
 
-export default function QuestionComponent({question, addAnswer}) {
+export default function QuestionComponent({question}) {
 	const [toggle, setToggle] = useState(true);
+	const questions = useStore(state => state.questions);
+	const addNewAnswer = useStore(state => state.addNewAnswer);
+
+	const addAnswer = (event, id) => {
+		event.preventDefault();
+		const newQuestions = questions.map(question => {
+			if (id === question.id) {
+				const newAnswer = {
+					id: nanoid(),
+					answerText: event.target.inputField.value,
+					answers: [],
+				};
+				question.answers = [...question.answers, newAnswer];
+			}
+			return question;
+		});
+		addNewAnswer(newQuestions);
+	};
 
 	return (
 		<StyledQuestionComponent>
