+++ src/hooks/useStore.js
@@ -1,16 +1,21 @@
 import create from 'zustand';
 
-const useStore = create(set => ({
-	counter: 0,
-	setCounter(counter) {
-		set({counter});
+import questionDatabase from '../utils/db';
+
+export const useStore = create(set => ({
+	questions: [...questionDatabase],
+	addNewAnswer: newQuestion => {
+		set(() => {
+			return {
+				questions: newQuestion,
+			};
+		});
 	},
-	decrementCounter(step = 1) {
-		set(({counter}) => ({counter: counter - step}));
-	},
-	incrementCounter(step = 1) {
-		set(({counter}) => ({counter: counter + step}));
+	addNewQuestion: newQuestion => {
+		set(state => {
+			return {
+				questions: [newQuestion, ...state.questions],
+			};
+		});
 	},
 }));
-
-export default useStore;
