+++ src/pages/questions.js
@@ -1,46 +1,11 @@
-import {nanoid} from 'nanoid';
-import {useState} from 'react';
 import {Helmet} from 'react-helmet';
 
 import Layout from '../components/Layout';
 import QuestionComponent from '../components/QuestionComponent';
+import {useStore} from '../hooks/useStore';
 
 export default function QuestionPage() {
-	const [questions, setQuestions] = useState([
-		{
-			id: nanoid(),
-			questionText: 'What is export default?',
-			answers: [
-				{id: nanoid(), answerText: 'its ...'},
-				{id: nanoid(), answerText: 'it means that ...'},
-			],
-		},
-		{
-			id: nanoid(),
-			questionText: 'What does clean code means?',
-			answers: [{id: nanoid(), answerText: 'its ...'}],
-		},
-		{
-			id: nanoid(),
-			questionText: 'Why do we have to define keys?',
-			answers: [],
-		},
-	]);
-
-	const addAnswer = (event, id) => {
-		event.preventDefault();
-		const index = questions.findIndex(question => {
-			return id === question.id;
-		});
-		const newAnswer = {
-			id: nanoid(),
-			answerText: event.target.inputField.value,
-		};
-		const newQuestions = [...questions];
-		newQuestions[index].answers = [...newQuestions[index].answers, newAnswer];
-		setQuestions(newQuestions);
-		//setQuestions(newQuestions);
-	};
+	const questions = useStore(state => state.questions);
 
 	return (
 		<Layout>
@@ -54,9 +19,7 @@ export default function QuestionPage() {
 			</Helmet>
 			<article>
 				{questions.map(question => {
-					return (
-						<QuestionComponent key={nanoid} question={question} addAnswer={addAnswer} />
-					);
+					return <QuestionComponent key={question.id} question={question} />;
 				})}
 			</article>
 		</Layout>
